# internal/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Vozzera API
  description: |
    API do Vozzera — salas de texto e voz, mensagens, autenticação e presença em tempo real.
    WebSocket em `/api/ws` com protocolo versionado (envelope `v:1`).
    Presença de voz via LiveKit Cloud com eventos replicados para o WebSocket.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local
security:
  - cookieAuth: []
...
paths:
...
  /api/voice/token:
    post:
      tags: [voice]
      summary: Gerar token do LiveKit para entrar em sala de voz
      operationId: getVoiceToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [room_id]
              properties:
                room_id: { type: string, format: uuid }
      responses:
        "200":
          description: Token e URL do LiveKit
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  url: { type: string }
                  room_name: { type: string }
        "400":
          description: Sala não é de voz ou payload inválido
        "401":
          description: Não autenticado
        "404":
          description: Sala não encontrada
  /api/voice/webhook:
    post:
      tags: [voice]
      summary: Webhook do LiveKit Cloud
      operationId: voiceWebhook
      description: |
        Recebe eventos de presença do LiveKit Cloud (participante entrou/saiu, sala encerrada).
        Validado pela assinatura HMAC do LiveKit. Não usa cookie de sessão.
        Os eventos são replicados para os clientes WebSocket conectados.
      security: []
      responses:
        "200":
          description: Evento processado
        "400":
          description: Assinatura inválida ou payload malformado
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: Cookie HttpOnly com o id opaco da sessão, definido em POST /api/login. Sessão revogável via POST /api/logout.
...